chore(Datamapper): Split Source and Target documents components
As preparation for https://github.com/KaotoIO/kaoto/issues/2396, we need
to split the Source and Target document's component to transition to a
lazy document parsing and generation in order to separate it from the
rendering logic.

relates: https://github.com/KaotoIO/kaoto/issues/2396
